airflow_scheduler/requirements.txt
airflow-provider-fivetran-async==1.0.0a4